LayoutTests/platform/chromium-mac/editing/selection/caret-mode-paragraph-keys-navigation-expected.txt

This tests that arrow keys navigate through a paragraph as expected when in caret browsing mode, also with shift and ctrl modifiers.

This tests that keys navigation through a paragraph work as expected when in caret mode, also with shift and ctrl modifiers.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS getSelection().baseOffset is 15
PASS getSelection().baseOffset is 16
PASS getSelection().baseOffset is 15
PASS getSelection().baseOffset is 15
PASS getSelection().extentOffset is 18
PASS getSelection().baseOffset is 15
PASS getSelection().extentOffset is 17
FAIL getSelection().baseOffset should be 21. Was 15.
FAIL getSelection().baseOffset should be 16. Was 15.
FAIL getSelection().baseOffset should be 16. Was 15.
FAIL getSelection().extentOffset should be 26. Was 17.
FAIL getSelection().baseOffset should be 22. Was 15.
PASS successfullyParsed is true

TEST COMPLETE
